feat: Make it work as a local packaged .exe
