+++ .github/workflows/build-and-release.yml
@@ -1,0 +1,32 @@
+name: Build and Release
+on:
+  push:
+    tags:
+      - v*
+
+jobs:
+  build:
+    runs-on: windows-latest
+    if: github.actor != 'dependabot[bot]'
+
+    steps:
+      - uses: actions/checkout@v3
+      - name: Use Node.js v16
+        uses: actions/setup-node@v3
+        with:
+          node-version: 16
+          cache: "npm"
+      - name: Install dependencies
+        run: npm ci
+      - name: Build
+        run: npm run build
+      - name: Archive Release
+        uses: thedoctor0/zip-release@master
+        with:
+          type: "zip"
+          directory: "./dist"
+          filename: "muk.zip"
+      - uses: ncipollo/release-action@v1
+        with:
+          artifacts: "./dist/muk.zip"
+          token: ${{ secrets.GITHUB_TOKEN }}
