+++ .gitignore
@@ -1,3 +1,4 @@
 .ssh
 .vscode
 node_modules
+dist/*.exe
