--- .lando.yml
@@ -1,16 +1,0 @@
-name: mukchests
-services:
-  node:
-    type: node:16
-    port: 3000
-    build:
-      - npm install
-    command: npm start
-tooling:
-  npm:
-    service: node
-  node:
-    service: node
-proxy:
-  node:
-    - mukchests.lndo.site:3000
