--- index.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Chest Troll 1.0</title>
-    <link rel="stylesheet" href="css/style.css">
-    <script src="https://code.jquery.com/jquery-3.6.0.slim.min.js" integrity="sha256-u7e5khyithlIdTpu22PHhENmPcRdFiHRjhAuHcs05RI=" crossorigin="anonymous"></script>
-  </head>
-  <body>
-    <div id="container"></div>
-    <div id="controls">
-      <input title="Batch #" type="number" value="0" name="batch" id="batch" disabled>
-      <input title="Total #" type="number" value="0" name="count" id="count" disabled>
-      <button id="add">Add</button>
-      <button id="remove">Remove</button>
-      <button id="clear">Clear</button>
-      <br>
-      <label for="number">Amount: </label>
-      <input type="number" value="1" name="number" id="number" step="1" min="1">
-      <label for="timer">Minutes: </label>
-      <input type="number" value="0" name="timer" id="timer" step="1" min="0">
-      <label for="timer">Random: </label>
-      <input type="checkbox" name="random" id="random">
-    </div>
-    <script type="text/javascript" src="js/logic.js"></script>
-    <script type="text/javascript" src="js/websocket.js"></script>
-  </body>
-</html>
