+++ src/public/css/style.css
@@ -1,0 +1,174 @@
+html, body {
+  background: rgba(0, 0, 0, 0);
+
+  padding: 0;
+  margin: 0;
+  height: 100%;
+
+  font-family: sans-serif;
+  font-size: 18px;
+  line-height: 1;
+
+  overflow: hidden;
+  position: relative;
+}
+
+body {
+  background-color: rgba(255,255,255,0.5);
+}
+
+#controls {
+  background-color: rgba(255,255,255,0.5);
+  display: none;
+  left: 1em;
+  line-height: 1.5em;
+  padding: 0 .5em;
+  position: absolute;
+  top: 1em;
+  z-index: 100;
+}
+
+button {
+  cursor: pointer;
+}
+
+input[type=number] {
+  width: 3em;
+}
+
+label {
+  font-size: 0.8rem;
+}
+
+#container {
+  display: flex;
+
+  flex-wrap: wrap;
+  align-items: flex-end;
+  align-content: flex-end;
+  justify-content: flex-end;
+  gap: 1em;
+
+  width: auto;
+  height: 100%;
+  position: relative;
+
+  right: 0;
+  bottom: 1em;
+
+  padding: 0 1em;
+}
+
+.chest-wrapper {
+  border-radius: 25%;
+  cursor: pointer;
+  font-size: 12px;
+  margin-left: -0.75em;
+  min-height: 7em;
+  min-width: 7em;
+  position: relative;
+}
+
+.chest-wrapper.random {
+  position: absolute;
+}
+
+.chest,
+.sparkle-cw,
+.sparkle-ccw {
+  background-repeat: no-repeat;
+  background-size: contain;
+  height: 100%;
+  position: absolute;
+  width: 100%;
+}
+
+.chest {
+  animation-name: wobble;
+  animation-delay: 2s;
+  animation-duration: 2.5s;
+  animation-iteration-count: infinite;
+  background-image: url("../images/gw2-chest.png");
+  height: 80%;
+  left: 10%;
+  width: 80%;
+  top: 15%;
+  z-index: 4;
+}
+
+.sparkle-ccw {
+  animation: ccwspin 10s linear infinite;
+  background-image: url("../images/ccw-sparkle.png");
+  margin-top: -0.5em;
+  opacity: 0.75;
+  z-index: 1;
+}
+
+.sparkle-cw {
+  animation: cwspin 5s linear infinite;
+  background-image: url("../images/cw-sparkle.png");
+  z-index: 2;
+}
+
+.chest:hover {
+  filter: brightness(0.75) ;
+}
+
+.chest:before {
+  color: #fff;
+  display: inline-block;
+  padding: 0.25em;
+}
+
+@keyframes ccwspin {
+  100% {
+      transform:rotate(-360deg);
+  }
+}
+
+@keyframes cwspin {
+  100% {
+      transform:rotate(360deg);
+  }
+}
+
+@keyframes wobble {
+  0% { transform: rotate(0deg); }
+  65% { transform: rotate(0deg); }
+  70% { transform: rotate(-20deg); }
+  75% { transform: rotate(20deg); }
+  80% { transform: rotate(-20deg); }
+  85% { transform: rotate(20deg); }
+  90% { transform: rotate(-10deg); }
+  95% { transform: rotate(10deg); }
+}
+
+/* Debugging (add 'debug' class to HTML element) */
+.debug {
+  background-image: url("../images/gw2-screenshot.jpg");
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-position: 100%;
+}
+
+.debug #controls {
+  display: block;
+}
+
+.debug .chest-wrapper::after {
+  content: attr(id);
+  font-size: 10px;
+  display: block;
+  text-align: center;
+  width: 100%;
+}
+
+/* Timer chests */
+.debug .chest-wrapper {
+  color: blue;
+}
+
+/* Normal chests */
+.debug .chest-wrapper[id$="-0"] {
+  color: black;
+}
